# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "open-cam API",
    "description": "open-cam 是视频流分析与监控管理平台。\n\n- **摄像头**：接入局域网 RTSP 流或本地视频文件，YOLO（cuda/mps/cpu 自适应）检测与跟踪。\n- **规则引擎**：区域入侵 / 徘徊逗留 / 人数统计 / 区域人数 / 越线计数，支持生效时段与冷却去抖。\n- **事件**：命中规则即落库并存快照，可经 VLM（OpenAI 兼容接口）异步复核；处置闭环支持关注星标、负责人指派、状态流转与 webhook 通知，全程留痕。\n- **方案包**：行业规则模板包（连锁零售 / 美容美发 / 餐饮 / 快餐），一键安装与应用。\n- **统计**：分时段进出店客流等聚合视图。\n\n交互式调试见 [Swagger UI](/docs)，结构化阅读见 [ReDoc](/redoc)。\n",
    "contact": {
      "name": "open-cam",
      "url": "https://github.com/local/open-cam"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "事件列表",
        "description": "支持 camera_id / rule_type / vlm_verdict / verdict / acked / status / starred / needs_action 过滤与 limit/offset 分页，按时间倒序。不传 needs_action 返回全部。",
        "operationId": "list_events_api_events_get",
        "parameters": [
          {
            "name": "camera_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按摄像头过滤",
              "title": "Camera Id"
            },
            "description": "按摄像头过滤"
          },
          {
            "name": "rule_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按规则类型过滤，如 zone_intrusion / line_crossing",
              "title": "Rule Type"
            },
            "description": "按规则类型过滤，如 zone_intrusion / line_crossing"
          },
          {
            "name": "vlm_verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按 VLM 判定过滤：confirmed / false_alarm / uncertain",
              "title": "Vlm Verdict"
            },
            "description": "按 VLM 判定过滤：confirmed / false_alarm / uncertain"
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按人工判定过滤：confirmed / false_alarm / unclear",
              "title": "Verdict"
            },
            "description": "按人工判定过滤：confirmed / false_alarm / unclear"
          },
          {
            "name": "acked",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按确认状态过滤",
              "title": "Acked"
            },
            "description": "按确认状态过滤"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按处置状态过滤：open / acked / resolved / ignored / logged",
              "title": "Status"
            },
            "description": "按处置状态过滤：open / acked / resolved / ignored / logged"
          },
          {
            "name": "starred",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "仅看关注（星标）事件",
              "title": "Starred"
            },
            "description": "仅看关注（星标）事件"
          },
          {
            "name": "needs_action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true=待办，false=观察记录；不传返回全部",
              "title": "Needs Action"
            },
            "description": "true=待办，false=观察记录；不传返回全部"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "每页条数",
              "default": 50,
              "title": "Limit"
            },
            "description": "每页条数"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "分页偏移",
              "default": 0,
              "title": "Offset"
            },
            "description": "分页偏移"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response List Events Api Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "事件详情",
        "description": "含命中目标、快照路径、VLM 复核结论与处置状态。",
        "operationId": "get_event_api_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "events"
        ],
        "summary": "编辑处置信息",
        "description": "更新判定/状态/星标/负责人/备注；观察记录不可改处置字段。",
        "operationId": "update_event_api_events__event_id__patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}/actions": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "处置时间线",
        "description": "该事件的全部处置记录（关注/指派/状态/备注/通知），按时间升序。",
        "operationId": "list_actions_api_events__event_id__actions_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventActionOut"
                  },
                  "title": "Response List Actions Api Events  Event Id  Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}/ack": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "确认事件（ack）",
        "operationId": "ack_event_api_events__event_id__ack_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}/notify": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "重发通知",
        "description": "把该事件重新推送到所有匹配的通知渠道，返回推送渠道数。",
        "operationId": "resend_notify_api_events__event_id__notify_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}/snapshot": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "事件快照图（JPEG）",
        "operationId": "event_snapshot_api_events__event_id__snapshot_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}/clip": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "事件对应的素材片段",
        "description": "文件源事件按命中点前后窗口回放；优先返回抽出的短 mp4，否则回源文件。",
        "operationId": "event_clip_api_events__event_id__clip_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}/feedback": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "误报/漏报反馈并写入训练数据集",
        "operationId": "event_feedback_api_events__event_id__feedback_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventFeedback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventActionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "ts": {
            "type": "number",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_id",
          "action",
          "actor",
          "payload",
          "ts"
        ],
        "title": "EventActionOut"
      },
      "EventFeedback": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "归入的训练任务 id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "false_alarm 误报 / miss 漏报"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "覆盖默认类别"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "kind"
        ],
        "title": "EventFeedback"
      },
      "EventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "camera_id": {
            "type": "integer",
            "title": "Camera Id"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "ts": {
            "type": "number",
            "title": "Ts"
          },
          "snapshot_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Path"
          },
          "source_offset": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Offset"
          },
          "camera_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Name"
          },
          "source_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Filename"
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          },
          "vlm_status": {
            "type": "string",
            "title": "Vlm Status"
          },
          "vlm_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlm Verdict"
          },
          "vlm_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlm Reason"
          },
          "acked": {
            "type": "boolean",
            "title": "Acked"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "starred": {
            "type": "boolean",
            "title": "Starred"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "default": "alert"
          },
          "needs_action": {
            "type": "boolean",
            "title": "Needs Action",
            "default": true
          },
          "repeat_count": {
            "type": "integer",
            "title": "Repeat Count",
            "default": 1
          },
          "clip_start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clip Start",
            "readOnly": true
          },
          "clip_end": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clip End",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "camera_id",
          "rule_id",
          "type",
          "confidence",
          "ts",
          "snapshot_path",
          "detail",
          "vlm_status",
          "vlm_verdict",
          "vlm_reason",
          "acked",
          "status",
          "starred",
          "assignee",
          "note",
          "clip_start",
          "clip_end"
        ],
        "title": "EventOut"
      },
...
      "EventUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(open|acked|resolved|ignored)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "starred": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starred"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(confirmed|false_alarm|unclear)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "EventUpdate",
        "description": "处置编辑：全部可选，只更新传入的字段。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}